feat(subsystem-4): Task 8 follow-up GREEN — float median + RED tests for round-boundary bias
Surgical 2-line code fix implementing the spec section 5.2 amendment
(committed aa1fb47) that resolves the round-boundary bias bug surfaced
by the joint cockpit retrospective.

Code changes (src/dynasty_genius/eval/backtest_mock_draft.py):
- ProspectConsensus.projected_pick_median: Optional[int] -> Optional[float]
  (half-picks preserved; no truncation; inline comment references the
  spec section 5.2 amendment)
- aggregate_per_prospect: median_val = int(_median(pick_nos)) ->
  float(_median(pick_nos)). Round-bucket assignment uses round_half_up
  at Task 10/12 bucket-mapping time (not here).

Test additions (tests/contract/test_subsystem_4_aggregation.py, Codex
re-authored RED):
- test_aggregation_median_preserves_float_and_rounding: 4 distinct
  sources, pick_nos [31, 32, 33, 34], round_tier_only tier. Asserts
  consensus.projected_pick_median == 32.5 AND isinstance float. RED on
  pre-GREEN impl (returns 32); GREEN passes.
- test_aggregation_iqr_uses_exclusive_quantiles: 4 distinct sources,
  pick_nos [10, 20, 30, 40], round_tier_only tier. Asserts
  consensus.projected_pick_iqr == 25.0 (exclusive method). Locks the
  aggregation_version IQR-method contract. Already passes on existing
  impl (Python statistics.quantiles default is exclusive); contract
  test prevents regression.

Cockpit cycle (unanimous):
- Codex authored RED, verified RED state pre-GREEN (median test FAIL,
  IQR test PASS as expected); reviewed GREEN diff and confirmed
  technical PASS
- Gemini reviewed GREEN governance attestation PASS

Test results: tests/contract/test_subsystem_4_aggregation.py -> 11/11
PASS; S4 contract guard (schema + bridge + ingestion + aggregation) ->
71/71 PASS; ruff clean; full suite (excluding obsoleted untracked
test_subsystem_4_join.py) -> 1447 passed, 11 skipped, 0 failed (up from
baseline 1445 by the 2 new tests).

Backward compatibility: existing tests that assert
projected_pick_median == 13 still pass because 13.0 == 13 in Python.
No test edits needed for "easy median" cases.

Branch: feature/subsystem-4-backtest-harness
Next: AGENT_SYNC update + ledger session-close + delete obsoleted
tests/contract/test_subsystem_4_join.py (David authorized) + push branch.

+++ src/dynasty_genius/eval/backtest_mock_draft.py
@@ -423,7 +423,7 @@ class ProspectConsensus(BaseModel):
     model_config = ConfigDict(extra="forbid")
 
     prospect_uuid: str
-    projected_pick_median: Optional[int] = None
+    projected_pick_median: Optional[float] = None  # half-picks preserved (no truncation); spec §5.2 amended 2026-05-28
     projected_pick_iqr: Optional[float] = None
     projected_pick_min: Optional[int] = None
     projected_pick_max: Optional[int] = None
@@ -476,7 +476,7 @@ def aggregate_per_prospect(
         staleness = _whole_day_staleness(draft_date, most_recent_published)
 
         pick_nos = [p.pick_no for p in picks]
-        median_val = int(_median(pick_nos))
+        median_val = float(_median(pick_nos))
         min_val = min(pick_nos)
         max_val = max(pick_nos)
         if len(pick_nos) >= 2:
